pyproject.toml
[tool.poetry]
name = "blackjack"
version = "0.1.0"
description = ""
authors = [ "Denise Nanni <denise.nanni1@gmail.com>" ]
readme = "README.md"
packages = [
    { include = "database_interface", from = "main_server" },
    { include = "client_interface", from = "main_server" },
    { include = "servers_interface", from = "main_server" },
    { include = "src", from = "game_server" },
    { include = "database", from = "main_server/database_interface" },
    { include = "client", from = "main_server/client_interface" },
    { include = "servers", from = "main_server/servers_interface" }
]

[tool.poetry.dependencies]
python = ">=3.11"
sqlalchemy = ">=2.0.40,<3.0.0"
psycopg2 = ">=2.9.10,<3.0.0"
flask = ">=3.1.0,<4.0.0"
flask-socketio = ">=5.5.1,<6.0.0"
requests = "^2.32.3"
flask-login = "^0.6.3"
python-dotenv = "^1.1.0"
cryptography = "^45.0.2"
pyjwt = "^2.10.1"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
database = "database_interface.database.main:main"
client = "client_interface.client.main:main"
servers = "servers_interface.servers.main:main"
game_server = "game_server.src.main:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

